src/backend/services/crewai-agents/requirements.txt
# CrewAI Production Deployment - Dependency Resolution
# Strategy: Use latest stable versions where compatibility issues are resolved

# Core Flask and API dependencies
flask>=2.0.0
flask-cors>=3.0.0
python-dotenv>=0.19.0

# Web scraping and HTTP requests  
requests>=2.25.0
urllib3>=1.26.0
beautifulsoup4>=4.10.0
lxml[html_clean]>=4.9.0
feedparser>=6.0.0

# Social media scraping dependencies
praw>=7.0.0  # Reddit API wrapper
python-telegram-bot>=20.0  # Telegram Bot API

# News extraction and parsing
newspaper3k>=0.2.8

# CrewAI and AI dependencies - LATEST STABLE APPROACH
# Use latest CrewAI and let it manage its own dependencies
crewai[tools]

# AI API clients - no version constraints to avoid conflicts
anthropic>=0.7.0

# Advanced scraping tools
firecrawl-py>=0.0.8
selenium>=4.0.0

# Utility dependencies  
python-dateutil>=2.8.0
pytz>=2021.3

# YAML support
pyyaml>=5.1

# Type hints and validation
pydantic>=2.0.0
typing-extensions>=4.0.0

# DEPLOYMENT STRATEGY:
# 1. Use latest CrewAI (should have ResponseTextConfig issue resolved)
# 2. Let pip resolve dependencies automatically
# 3. If issues persist, implement runtime compatibility patches